% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc.stat.R
\name{calc.stat}
\alias{calc.stat}
\title{Calculates a set of statistics for a given array of numbers.}
\usage{
calc.stat(x)
}
\arguments{
\item{x}{a numeric array}
}
\value{
An array of calculated statistics of a given set of numbers
}
\description{
\pkg{calc.stat} calculates a set of statistics for a given array of numbers.
}
\details{
\pkg{calc.stat} calculates a set of statistics for a given array of numbers.
The calculated statistics are
\itemize{
 \item Ntot = length of the array
 \item Nunique = Number of unique elements
 \item Min = Minimum value of the array
 \item Max = Maximum value of the array
 \item Mean = Mean value of the array
 \item Median = Median value of the array
 \item SD = Standard deviation value of the array
 \item SE = Standard error value of the array
 \item CVp = Percent coefficient of variation of the array
 \item CI95u = Upper limit of the 95\% confidence interval of the array
 \item cI95l = Lower limit of the 95\% confidence interval of the array
 \item gMean = Geometric mean value of the array
 \item gCVp = Geometric percent coefficient of variation of the array
}
}

