% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dv.plot.R
\name{dv.plot}
\alias{dv.plot}
\title{Plots drug plasma concentration vs time data}
\usage{
dv.plot(pdata, cunit = "[M].[L]^-3", tunit = "[T]")
}
\arguments{
\item{pdata}{A data frame with four mandatory columns:
\enumerate{
 \item Time: Time column
 \item Conc: DV column
 \item ID: Individual ID
 \item FCT: Dose identifier and/or population stratifier
}}

\item{cunit}{Unit for concentration (\strong{"[M].[L]^-3"})}

\item{tunit}{Unit for time (\strong{"[T]"})}
}
\value{
returns a graphical object created by arrangeGrob function
}
\description{
\pkg{dv.plot} plots DV vs Time data.
}
\details{
\pkg{dv.plot} plots DV vs Time data
}

