% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.read.sim.R
\name{nca.read.sim}
\alias{nca.read.sim}
\title{Check observed data}
\usage{
nca.read.sim(simFile = "nca_simulation.1.npctab.dta", MDV.rm = FALSE)
}
\arguments{
\item{simFile}{NONMEM simulation output with the simulated
concentration-time data from an internal data frame or an external table.
Default is \strong{"nca_simulation.1.npctab.dta"}.}

\item{MDV.rm}{If \code{TRUE} MDV column is used to remove non-observation
rows. Default is \strong{\code{FALSE}}}
}
\value{
A list of objects
}
\description{
\pkg{nca.read.sim} Reads NONMEM simulation output file.
}
\details{
\pkg{nca.read.sim} Reads NONMEM simulation output file.
}

