% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.digits.R
\name{out.digits}
\alias{out.digits}
\title{output value with correct digits and trailing zero}
\usage{
out.digits(x, dig = 3)
}
\arguments{
\item{x}{is the value}

\item{dig}{is the number of significant digits}
}
\description{
Function to present a value with correct digits and trailing zero
}
\details{
This is a function to present a value with correct digits and trailing zero. Numbers >= 10000, or <= 0.0001 will be presented in scientific format
}
\examples{
\dontrun{
out.digits(1234)
}

}
