% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_nm_sim.R
\name{read_nm_sim}
\alias{read_nm_sim}
\title{Read nonmem table files produced from simulation.}
\usage{
read_nm_sim(table_sim, only_obs = TRUE)
}
\arguments{
\item{table_sim}{The simulated table file to read. A text string.}

\item{only_obs}{Should the non-observation lines in the data set be removed?
  Currently filtered uisng the expected \code{MDV} column. \code{TRUE} or
  \code{FALSE}.}
}
\value{
Returns a data frame of the simulated table with an added column for
  the simulation number. The data frame is given class \code{c("tbl_df",
  "tbl", "data.frame")} for easy use with \code{\link[dplyr]{dplyr}}.
}
\description{
The function reads in nonmem table files produced from the \code{$SIM} line
in a NONMEM model file.
}
\details{
Currently the function expects the \code{$TABLE} to have a header for each
new simulation. This means that the \code{NOHEADER} option or
\code{ONEHEADER} option in the table file is not allowed.
}

