% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dv.plot.R
\name{dv.plot}
\alias{dv.plot}
\title{Plots drug plasma concentration vs time data}
\usage{
dv.plot(pdata, cunit = "[M].[L]^-3", tunit = "[T]")
}
\arguments{
\item{pdata}{A data frame with three mandatory columns and one optional
  column:
\enumerate{
 \item Time: Time column (mandatory)
 \item Conc: DV column (mandatory)
 \item ID: Individual ID (mandatory)
 \item FCT: Dose identifier and/or population stratifier (optional)
}}

\item{cunit}{Unit for concentration (\strong{"[M].[L]^-3"})}

\item{tunit}{Unit for time (\strong{"[T]"})}
}
\value{
returns a graphical object created by arrangeGrob function
}
\description{
\pkg{dv.plot} plots DV vs Time data.
}
\details{
\pkg{dv.plot} plots DV vs Time data
}

